move header test case to head component folder
