+++ .travis.yml
@@ -8,7 +8,7 @@ install:
   - yarn install
 
 script:
-  - yarn test
+  - yarn lint:test
   - yarn run build
 
 deploy:
