+++ frontend/jest.config.js
@@ -1,4 +1,7 @@
 module.exports = {
   setupFilesAfterEnv: ["<rootDir>src/test/setupTests.js"],
   snapshotSerializers: ["enzyme-to-json/serializer"],
+  moduleNameMapper: {
+    "^.+\\.(css|less|scss)$": "babel-jest",
+  },
 };
