+++ frontend/package.json
@@ -53,6 +53,7 @@
     "eslint-plugin-react-hooks": "^4.1.0",
     "html-loader": "1.1.0",
     "html-webpack-plugin": "^3.2.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^26.4.2",
     "mini-css-extract-plugin": "^0.10.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
