+++ frontend/src/assets/sass/_main.scss
@@ -10,4 +10,5 @@ body {
 
 * {
   box-sizing: border-box;
+  margin: 0;
 }
