+++ frontend/src/assets/sass/styles.scss
@@ -4,8 +4,6 @@
 
 @import "main";
 
-@import "navigation";
-
 @import "components";
 
 @import "media-queries";
