+++ frontend/src/js/components/header/_header.scss
@@ -1,0 +1,11 @@
+@import "../../../assets/sass/styles.scss";
+
+.nav__bar__container {
+  display: flex;
+  width: 100vw;
+  background-color: $black;
+
+  .nav__links__container {
+    color: $white;
+  }
+}
