+++ frontend/src/js/containers/AppContainer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Header from "../components/Header";
+import Header from "../components/header/Header";
 import { MainContext } from "../utils/contextUtils";
 
 const AppContainer = ({ children }) => {
